Change errors found while processing inputs to not raise an exception
This is because even though the contract is not being fulfilled, I believe
that the called code also has a contract to behave in a certain well-defined
way. That is, to return a failure with an explanation of why it can't do
it's specified task. Ideally, we would check for this thing at compile time
but since ruby is an interpreted language, our options are limited.